runpod_worker/requirements.txt
# Core
runpod==1.6.2
requests>=2.31.0
pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
matplotlib>=3.7.0
scikit-image>=0.21.0
opencv-python-headless>=4.8.0.74

# PyTorch pinned to image CUDA toolchain (installed in base image)
# torch
# torchvision

# ESRGAN
basicsr==1.4.2
realesrgan==0.3.0

# YOLOv5 deps
PyYAML>=6.0
pandas>=2.0.0
seaborn>=0.13.0
tqdm>=4.66.0
